% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3prop_utils.R
\name{AUROC}
\alias{AUROC}
\title{Compute area under the ROC curve}
\usage{
AUROC(p, y, plot_roc = TRUE, warn_inv = TRUE)
}
\arguments{
\item{p}{vector of scores. Can be sparse (from package Matrix), but will be
coerced to dense.}

\item{y}{vector of labels (0 or 1). Can be sparse (from package Matrix), but
will be coerced to dense.}

\item{plot_roc}{if \code{TRUE}, plots the ROC curve.}

\item{warn_inv}{raises a warning if high values of \code{p} are associated
with the 0 label, which produces an AUROC lower than 0.5.}
}
\value{
the value of the AUROC.
}
\description{
Function for computing the area under the ROC curve (AUROC)
}
\details{
It is assumed that a higher score increases the probability of an object
having label 1.
}
