% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dta}
\alias{example_dta}
\title{Example dataset}
\format{A list with:
\describe{
  \item{net}{the affinity matrix}
  \item{labels}{the labels}
}}
\source{
\url{http://ai.stanford.edu/~saram/3propcode.zip}
}
\usage{
example_dta
}
\description{
A dataset containing an affinity matrix and a vector of labels. It is part of
the code provided by the authors.
}
\keyword{datasets}
