% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3prop_normalize_A.R
\name{normalize_A}
\alias{normalize_A}
\title{Normalize an affinity matrix}
\usage{
normalize_A(A, M_type)
}
\arguments{
\item{A}{a symmetric affinity matrix. Can be dense or sparse from package
\code{Matrix}.}

\item{M_type}{a character string with one of "asym", "sym", or "as_is" (see
details).}
}
\value{
A sparse normalized affinity matrix.
}
\description{
This function provides different modes of normalizing an affinity matrix.
}
\details{
All methods for normalization start by removing possible diagonal elements.
The available methods are:
\describe{
    \item{asym}{asymmetric normalization, which yields a singly stochastic
    matrix, where rows add to 1.}
    \item{sym}{symmetric normalization, yielding a symmetric matrix.}
    \item{rm_diag}{only removes possible diagonal elements.}
}
}
\examples{
# simulate_simple_SBM does not produce self connections
A = as.matrix(simulate_simple_SBM(25L, 0.2, 0.001, 0.25)$A)
all.equal(normalize_A(A, "rm_diag"), A)

}
\references{
Mostafavi, S., Goldenberg, A., & Morris, Q. (2012). Labeling nodes using
three degrees of propagation. \emph{PloS one, 7}(12), e51947.
}
