% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_network.R
\name{simulate_simple_SBM}
\alias{simulate_simple_SBM}
\title{Simulate simplified Stochastic Block Model}
\usage{
simulate_simple_SBM(N, p_1, p_s, p_d, verb = FALSE)
}
\arguments{
\item{N}{number of nodes.}

\item{p_1}{probability of a node being labelled as 1.}

\item{p_s}{probability of an edge existing between nodes with equal labels.}

\item{p_d}{probability of an edge existing between nodes with different
labels.}
}
\value{
a named list with \code{y} being a sparse vector of labels, and
\code{A} a sparse adjacency matrix.
}
\description{
This function simulates a special case of the Stochastic Block Model (SBM).
}
\details{
In this simplified version, there are only 2 distinct labels. Nodes with
different labels have an edge between them with probability \code{p_d},
while the probability is \code{p_s} if they have matching labels. In the
language of the SBM, the matrix of edge probabilities is completely specified
by one number associated with the diagonal (\code{p_s}), and another for the
off-diagonal elements (\code{p_d}),

Note that this implementation takes advantage of the structure of the SBM
to efficiently sample sparse random objects. One consequence of this is that
the elements of \code{y} are ordered, starting with all the positive nodes,
and then all the non-positives. The same happens with \code{A}. If this is
an issue, the user can always permute the rows of the results.
}
\examples{
res = simulate_simple_SBM(25L, 0.2, 0.001, 0.25)
str(res)

}
\references{
Mostafavi, S., Goldenberg, A., & Morris, Q. (2012). Labeling nodes using
three degrees of propagation. \emph{PloS one, 7}(12), e51947.
}
